added HYPE dataset

+++ AI4Water/utils/datasets/datasets.py
@@ -63,9 +63,6 @@
 # https://zenodo.org/record/3836648#.YExu09yRWUk
 # https://zenodo.org/record/4290294#.YExo5tyRWUk
 # https://zenodo.org/record/2728636#.YEx4EdyRWUk
-# https://zenodo.org/record/581435#.YEx4KNyRWUk  HYPE
-# https://zenodo.org/record/4029572#.YEx5HNyRWUk  HYPE
-# https://zenodo.org/record/4029572#.YEx6Z9yRWUk  HYPE
 # https://zenodo.org/record/3581187#.YEx5CNyRWUk
 # https://zenodo.org/record/3946242#.YEx5FtyRWUk
 # https://zenodo.org/record/883100#.YEx5L9yRWUk
@@ -269,9 +266,14 @@ def sanity_check(dataset_name, path):
 
 class Datasets(object):
 
-    def __init__(self, name=None):
+    def __init__(self, name=None, units=None):
         if name is None:
             name = self.__class__.__name__
+
+        if units is not None:
+            assert units in ['si', 'imperial', 'metric']
+
+        self.units = units
         self.name = name
 
     @property
@@ -312,7 +314,7 @@ class Datasets(object):
 
     def _download(self, overwrite=False):
         """Downloads the dataset. If already downloaded, then"""
-        if os.path.exists(self.ds_dir):
+        if os.path.exists(self.ds_dir) and len(os.listdir(self.ds_dir))>0:
             if overwrite:
                 print(f"removing previous data directory {self.ds_dir} and downloading new")
                 shutil.rmtree(self.ds_dir)
@@ -328,9 +330,16 @@ Use overwrite=True to remove previously saved files and download again""")
         return
 
     def _download_and_unzip(self):
-        os.makedirs(self.ds_dir)
-        download_from_zenodo(self.ds_dir, self.url)
-        self._unzip()
+        if not os.path.exists(self.ds_dir):
+            os.makedirs(self.ds_dir)
+        if isinstance(self.url, str):
+            download_from_zenodo(self.ds_dir, self.url)
+            self._unzip()
+        elif isinstance(self.url, list):
+            for url in self.url:
+                download_from_zenodo(self.ds_dir, url)
+                self._unzip()
+
         return
 
     def _unzip(self, dirname=None):
@@ -595,7 +604,7 @@ when categorices is {categories}"""
 
 
 class LamaH(Camels):
-    f"""
+    """
     Large-Sample Data for Hydrology and Environmental Sciences for Central Europe
     from     url = "https://zenodo.org/record/4609826#.YFNp59zt02w"
     paper: https://essd.copernicus.org/preprints/essd-2021-72/
@@ -605,7 +614,7 @@ class LamaH(Camels):
 
     static_attribute_categories = ['']
 
-    def __init__(self, *, time_step, data_type, **kwargs):
+    def __init__(self, *, time_step:str, data_type:str, **kwargs):
         assert time_step in ['daily', 'hourly'], f"invalid time_step {time_step} given"
         assert data_type in self._data_types
         self.time_step = time_step
@@ -613,6 +622,12 @@ class LamaH(Camels):
         super().__init__(**kwargs)
 
         self._download()
+
+    """
+    Arguments:
+        time_step str:
+        data_type str:
+    """
 
     @property
     def dynamic_attributes(self):
@@ -688,7 +703,7 @@ class LamaH(Camels):
 
     def fetch_dynamic_attributes(self,
                                  station,
-                                 dynamic_attributes='all',
+                                 attributes='all',
                                  st=None,
                                  en=None,
                                  **kwargs):
@@ -698,7 +713,7 @@ class LamaH(Camels):
         if en is None:
             en = self.end
 
-        dynamic_attributes = check_attributes(dynamic_attributes, self.dynamic_attributes)
+        dynamic_attributes = check_attributes(attributes, self.dynamic_attributes)
 
         df = self.read_ts_of_station(station)
 
@@ -799,8 +814,10 @@ class HYSETS(Camels):
 
 
 class CAMELS_US(Camels):
-    """Downloads and processes CAMELS dataset of 671 catchments named as CAMELS
+    """
+    Downloads and processes CAMELS dataset of 671 catchments named as CAMELS
     from https://ral.ucar.edu/solutions/products/camels
+    https://doi.org/10.5194/hess-19-209-2015
     """
     DATASETS = ['CAMELS_US']
     url = "https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/basin_timeseries_v1p2_metForcing_obsFlow.zip"
@@ -1522,7 +1539,10 @@ class CAMELS_AUS(Camels):
 
 
 class CAMELS_CL(Camels):
-    """Downloads and processes CAMELS dataset of Chile"""
+    """
+    Downloads and processes CAMELS dataset of Chile
+    https://doi.org/10.5194/hess-22-5817-2018
+    """
 
     urls = {
         "1_CAMELScl_attributes.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
@@ -1714,6 +1734,75 @@ class CAMELS_CL(Camels):
             df = pd.read_csv(path, sep='\t', index_col='gauge_id', usecols=[station, 'gauge_id']).transpose()
 
         return df[attributes]
+
+
+class HYPE(Camels):
+    """
+    Downloads and preprocesses HYPE dataset from https://zenodo.org/record/4029572.
+    This is a rainfall-runoff dataset of 564 stations from 1985 to 2019 at daily
+    monthly and yearly time steps.
+    paper : https://doi.org/10.2166/nh.2010.007
+    """
+    url = [
+        "https://zenodo.org/record/581435",
+        "https://zenodo.org/record/4029572"
+    ]
+    dynamic_attributes = [
+        'AET_mm',
+        'Baseflow_mm',
+        'Infiltration_mm',
+        'SM_mm',
+        'Streamflow_mm',
+        'Runoff_mm',
+        'Qsim_m3-s',
+        'Prec_mm',
+        'PET_mm'
+    ]
+
+    def __init__(self, time_step:str='daily', **kwargs):
+        assert time_step in ['daily', 'monthly', 'yearly']
+        self.time_step = time_step
+        self.ds_dir = None
+        super().__init__(**kwargs)
+
+        self._download()
+
+    def stations(self):
+        return np.arange(1, 565).astype(str).tolist()
+
+    def fetch_dynamic_attributes(self,
+                                 station,
+                                 attributes='all',
+                                 st=None,
+                                 en=None,
+                                 **kwargs):
+
+        dynamic_attributes = check_attributes(attributes, self.dynamic_attributes)
+
+        _dynamic_attributes = []
+        for dyn_attr in dynamic_attributes:
+            pref, suff = dyn_attr.split('_')[0], dyn_attr.split('_')[-1]
+            _dyn_attr = f"{pref}_{self.time_step}_{suff}"
+            _dynamic_attributes.append(_dyn_attr)
+
+        df = pd.DataFrame()
+        for dyn_attr in _dynamic_attributes:
+            fname = f"{dyn_attr}.csv"
+            fpath = os.path.join(self.ds_dir, fname)
+            _df = pd.read_csv(fpath, index_col='DATE', usecols=['DATE', str(station)])
+            _df.index = pd.to_datetime(_df.index)
+            _df.columns = [dyn_attr]
+            df = pd.concat([df, _df], axis=1)
+
+        return df[_dynamic_attributes][st:en]
+
+    @property
+    def start(self):
+        return '19850101'
+
+    @property
+    def end(self):
+        return '20191231'
 
 
 class Weisssee(Datasets):
